refactor(core): モンスター一覧を battle.ts から分離する (Refs #419, #537)
モンスターのデータをソースコードに書くのは良くない (オーナー方針)。レコードが正で、
ソースは書く場所ではない。

一覧を monster-roster.ts へ分離し、ファイル先頭に「新しい敵・調整済みの値をここへ
書かない」を明記した。残っている理由はフォールバック (戦闘を止めない) とテスト・sim の
決定的な土台の 2 つで、レコード運用が安定したら中身を汎用プレースホルダへ縮める。

中身は無変更 (583 tests green で確認)。
