+++ docs/19-overworld.md
@@ -938,8 +938,12 @@ WorldPart { terrain: string; name: string; walkable?: boolean }
 `MONSTERS` はコード直書きで、リポジトリが公開なので**未実装の敵まで手の内が見える**。
 管理者 PDS の `world.monsters` レコードで差し替えられるようにした。
 
-- **フォールバック**: `レコード ?? コード直書き`。読めない/壊れていても今の敵で戦闘は続く。
-  ネタバレ解消は移行完了後にコードから消した時点で達成 (それまでの二重管理は許容)
+- **レコードが正。ソースは書く場所ではない** (オーナー方針)。同梱の一覧
+  (`monster-roster.ts`) はフォールバックとテストの土台であって、**新しい敵・調整済みの
+  値をそこへ書かない**。エディタで保存すればレコードが一覧を丸ごと差し替える
+- 移行手順: エディタを開く (同梱の一覧から始まる) → 保存 → 以後レコードが正。
+  レコード運用が安定したら同梱の中身を汎用プレースホルダへ縮める (#537 の最終段)。
+  読めない/壊れていてもフォールバックで戦闘は続く
 - **参照を保ったまま中身を差し替える**。`MONSTERS` / `MONSTERS_BY_ID` は 14 箇所以上から
   import されているので、呼び出し側は無変更。`MAX_POPULATED_TIER` は ESM の live binding
 - **edge も読む** (戦闘計算の権威)。web だけが編集後の敵を見ていると強さも XP も食い違う
